Change the inner class and update the travis
