+++ gradle.properties
@@ -25,7 +25,7 @@ slf4jApiVersion=1.7.25
 sparkVersion=3.0.1
 
 # Version and base tags can be overridden at build time.
-connectorVersion=0.8.0-SNAPSHOT
+connectorVersion=0.8.1-SNAPSHOT
 pravegaVersion=0.8.1
 
 # flag to indicate if Pravega sub-module should be used instead of the version defined in 'pravegaVersion'
