+++ src/main/scala/io/pravega/connectors/spark/PravegaSourceProvider.scala
@@ -11,29 +11,18 @@ package io.pravega.connectors.spark
 
 import java.net.URI
 import java.time.Duration
-import java.util
-import java.util.{Locale, Optional}
+import java.util.Locale
 
 import io.pravega.client.ClientConfig
 import io.pravega.client.admin.StreamManager
 import io.pravega.client.stream.{RetentionPolicy, ScalingPolicy, StreamConfiguration, StreamCut}
 import org.apache.spark.internal.Logging
-import org.apache.spark.sql.{AnalysisException, DataFrame, SQLContext, SaveMode}
-import org.apache.spark.sql.connector.catalog.{SupportsRead, SupportsWrite, Table, TableCapability, TableProvider}
-import org.apache.spark.sql.catalyst.InternalRow
-import org.apache.spark.sql.catalyst.expressions.GenericInternalRow
 import org.apache.spark.sql.catalyst.util.CaseInsensitiveMap
-import org.apache.spark.sql.connector.read.{Batch, Scan, ScanBuilder}
-import org.apache.spark.sql.connector.read.streaming.{ContinuousStream, MicroBatchStream}
-import org.apache.spark.sql.connector.write.streaming.StreamingWrite
-import org.apache.spark.sql.connector.write.{BatchWrite, LogicalWriteInfo, SupportsTruncate, WriteBuilder}
-import org.apache.spark.sql.execution.streaming.{Sink, Source}
-import org.apache.spark.sql.internal.connector.{SimpleTableProvider, SupportsStreamingUpdate}
+import org.apache.spark.sql.connector.catalog.Table
+import org.apache.spark.sql.internal.connector.SimpleTableProvider
 import org.apache.spark.sql.sources._
-import org.apache.spark.sql.util.CaseInsensitiveStringMap
-import org.apache.spark.sql.streaming.OutputMode
 import org.apache.spark.sql.types.{StructField, _}
-import org.apache.spark.unsafe.types.UTF8String
+import org.apache.spark.sql.util.CaseInsensitiveStringMap
 import resource.managed
 
 import scala.collection.JavaConverters._
@@ -49,10 +38,6 @@ class PravegaSourceProvider extends DataSourceRegister
   with SimpleTableProvider
   with Logging {
 
-  private val DEFAULT_CONTROLLER = "tcp://localhost:9090"
-    private val DEFAULT_TRANSACTION_TIMEOUT_MS: Long = 30 * 1000
-    private val DEFAULT_BATCH_TRANSACTION_TIMEOUT_MS: Long = 2 * 60 * 1000 // 2 minutes (maximum allowed by default server)
-    private val DEFAULT_TRANSACTION_STATUS_POLL_INTERVAL_MS: Long = 50
 
   /** String that represents the format that this data source provider uses. */
   override def shortName(): String = PravegaSourceProvider.SOURCE_PROVIDER_NAME
@@ -65,7 +50,7 @@ class PravegaSourceProvider extends DataSourceRegister
    */
   override def getTable(options: CaseInsensitiveStringMap): Table = {
     val caseInsensitiveParams = CaseInsensitiveMap(options.asScala.toMap)
-    validateBatchOptions(caseInsensitiveParams)
+    PravegaSourceProvider.validateOptions(caseInsensitiveParams)
     val metadataTableName = caseInsensitiveParams.get(PravegaSourceProvider.METADATA_OPTION_KEY).map(MetadataTableName.withName)
     if (metadataTableName.isDefined) {
       log.debug("return PravegaMetaTable")
@@ -76,9 +61,106 @@ class PravegaSourceProvider extends DataSourceRegister
     }
   }
 
+}
 
-  def validateStreamOptions(caseInsensitiveParams: Map[String, String]): Unit = {
-    validateGeneralOptions(caseInsensitiveParams)
+
+object PravegaSourceProvider extends Logging {
+  private[spark] val SOURCE_PROVIDER_NAME = "pravega"
+  private[spark] val CONTROLLER_OPTION_KEY = "controller"
+  private[spark] val SCOPE_OPTION_KEY = "scope"
+  private[spark] val STREAM_OPTION_KEY = "stream"
+  private[spark] val TRANSACTION_TIMEOUT_MS_OPTION_KEY = "transaction_timeout_ms"
+  private[spark] val START_STREAM_CUT_OPTION_KEY = "start_stream_cut"
+  private[spark] val END_STREAM_CUT_OPTION_KEY = "end_stream_cut"
+  private[spark] val ALLOW_CREATE_SCOPE_OPTION_KEY = "allow_create_scope"
+  private[spark] val ALLOW_CREATE_STREAM_OPTION_KEY = "allow_create_stream"
+  private[spark] val DEFAULT_NUM_SEGMENTS_OPTION_KEY = "default_num_segments"
+  private[spark] val READ_AFTER_WRITE_CONSISTENCY_OPTION_KEY = "read_after_write_consistency"
+  private[spark] val TRANSACTION_STATUS_POLL_INTERVAL_MS_OPTION_KEY = "transaction_status_poll_interval_ms"
+  private[spark] val METADATA_OPTION_KEY = "metadata"
+  private[spark] val DEFAULT_SCALE_FACTOR_OPTION_KEY = "default_scale_factor"
+  private[spark] val DEFAULT_SEGMENT_TARGET_RATE_BYTES_PER_SEC_OPTION_KEY = "default_segment_target_rate_bytes_per_sec"
+  private[spark] val DEFAULT_SEGMENT_TARGET_RATE_EVENTS_PER_SEC_OPTION_KEY = "default_segment_target_rate_events_per_sec"
+  private[spark] val DEFAULT_RETENTION_DURATION_MILLISECONDS_OPTION_KEY = "default_retention_duration_milliseconds"
+  private[spark] val DEFAULT_RETENTION_SIZE_BYTES_OPTION_KEY = "default_retention_size_bytes"
+  private[spark] val EXACTLY_ONCE = "exactly_once"
+
+  private[spark] val STREAM_CUT_EARLIEST = "earliest"
+  private[spark] val STREAM_CUT_LATEST = "latest"
+  private[spark] val STREAM_CUT_UNBOUNDED = "unbounded"
+  private[spark] val ROUTING_KEY_ATTRIBUTE_NAME = "routing_key"
+  private[spark] val EVENT_ATTRIBUTE_NAME = "event"
+
+  private[spark] val DEFAULT_CONTROLLER = "tcp://localhost:9090"
+  private[spark] val DEFAULT_TRANSACTION_TIMEOUT_MS: Long = 30 * 1000
+  private[spark] val DEFAULT_BATCH_TRANSACTION_TIMEOUT_MS: Long = 2 * 60 * 1000 // 2 minutes (maximum allowed by default server)
+  private[spark] val DEFAULT_TRANSACTION_STATUS_POLL_INTERVAL_MS: Long = 50
+
+
+  def buildStreamConfig(caseInsensitiveParams: Map[String, String]): StreamConfiguration = {
+    var streamConfig = StreamConfiguration.builder
+    val minSegments = caseInsensitiveParams.get(PravegaSourceProvider.DEFAULT_NUM_SEGMENTS_OPTION_KEY)
+    val scaleFactor = caseInsensitiveParams.get(PravegaSourceProvider.DEFAULT_SCALE_FACTOR_OPTION_KEY)
+    val targetRateBytesPerSec = caseInsensitiveParams.get(PravegaSourceProvider.DEFAULT_SEGMENT_TARGET_RATE_BYTES_PER_SEC_OPTION_KEY)
+    val targetRateEventsPerSec = caseInsensitiveParams.get(PravegaSourceProvider.DEFAULT_SEGMENT_TARGET_RATE_EVENTS_PER_SEC_OPTION_KEY)
+    val retentionDurationMilliseconds = caseInsensitiveParams.get(PravegaSourceProvider.DEFAULT_RETENTION_DURATION_MILLISECONDS_OPTION_KEY)
+    val retentionSizeBytes = caseInsensitiveParams.get(PravegaSourceProvider.DEFAULT_RETENTION_SIZE_BYTES_OPTION_KEY)
+
+    // set scaling policy
+    streamConfig = (minSegments, scaleFactor, targetRateBytesPerSec, targetRateEventsPerSec) match {
+      case (Some(minSegments), scaleFactor, targetRateKiloBytesPerSec, targetRateEventsPerSec) =>
+        (scaleFactor, targetRateKiloBytesPerSec, targetRateEventsPerSec) match {
+          case (Some(scaleFactor), Some(targetRateBytesPerSec), None) =>
+            streamConfig.scalingPolicy(ScalingPolicy.byDataRate(targetRateBytesPerSec.toInt / 1024, scaleFactor.toInt, minSegments.toInt))
+          case (Some(scaleFactor), None, Some(targetRateEventsPerSec)) =>
+            streamConfig.scalingPolicy(ScalingPolicy.byEventRate(targetRateEventsPerSec.toInt, scaleFactor.toInt, minSegments.toInt))
+          case _ =>
+            streamConfig.scalingPolicy(ScalingPolicy.fixed(minSegments.toInt))
+        }
+      case _ => streamConfig
+    }
+
+    // set retention policy
+    streamConfig = (retentionDurationMilliseconds, retentionSizeBytes) match {
+      case (Some(retentionDurationMilliseconds), None) =>
+        streamConfig.retentionPolicy(RetentionPolicy.byTime(Duration.ofMillis(retentionDurationMilliseconds.toLong)))
+      case (None, Some(retentionSizeBytes)) =>
+        streamConfig.retentionPolicy(RetentionPolicy.bySizeBytes(retentionSizeBytes.toLong))
+      case _ =>
+        streamConfig
+    }
+
+    log.info("streamConfig is {}", streamConfig)
+
+    return streamConfig.build()
+  }
+
+  def getPravegaStreamCut(
+                           params: Map[String, String],
+                           streamCutOptionKey: String,
+                           defaultStreamCut: PravegaStreamCut): PravegaStreamCut = {
+
+    params.get(streamCutOptionKey).map(_.trim) match {
+      case Some(offset) if offset.toLowerCase(Locale.ROOT) == STREAM_CUT_LATEST =>
+        LatestStreamCut
+      case Some(offset) if offset.toLowerCase(Locale.ROOT) == STREAM_CUT_EARLIEST =>
+        EarliestStreamCut
+      case Some(offset) if offset.toLowerCase(Locale.ROOT) == STREAM_CUT_UNBOUNDED =>
+        UnboundedStreamCut
+      case Some(base64String) => SpecificStreamCut(StreamCut.from(base64String))
+      case None => defaultStreamCut
+    }
+  }
+
+  def validateOptions(caseInsensitiveParams: Map[String, String]): Unit = {
+
+    // ValidateGeneralOptions
+    if (caseInsensitiveParams.getOrElse(PravegaSourceProvider.SCOPE_OPTION_KEY, "").isEmpty) {
+      throw new IllegalArgumentException(s"Missing required option '${PravegaSourceProvider.SCOPE_OPTION_KEY}'")
+    }
+    if (caseInsensitiveParams.getOrElse(PravegaSourceProvider.STREAM_OPTION_KEY, "").isEmpty) {
+      throw new IllegalArgumentException(s"Missing required option '${PravegaSourceProvider.STREAM_OPTION_KEY}'")
+    }
 
     if (caseInsensitiveParams.contains(PravegaSourceProvider.DEFAULT_SEGMENT_TARGET_RATE_BYTES_PER_SEC_OPTION_KEY) &&
       caseInsensitiveParams.contains(PravegaSourceProvider.DEFAULT_SEGMENT_TARGET_RATE_EVENTS_PER_SEC_OPTION_KEY)) {
@@ -121,28 +203,14 @@ class PravegaSourceProvider extends DataSourceRegister
     }
   }
 
-  private def validateBatchOptions(caseInsensitiveParams: Map[String, String]): Unit = {
-    // TODO: validate options
-    validateGeneralOptions(caseInsensitiveParams)
-  }
-
-  private def validateGeneralOptions(caseInsensitiveParams: Map[String, String]): Unit = {
-    if (caseInsensitiveParams.getOrElse(PravegaSourceProvider.SCOPE_OPTION_KEY, "").isEmpty) {
-      throw new IllegalArgumentException(s"Missing required option '${PravegaSourceProvider.SCOPE_OPTION_KEY}'")
-    }
-    if (caseInsensitiveParams.getOrElse(PravegaSourceProvider.STREAM_OPTION_KEY, "").isEmpty) {
-      throw new IllegalArgumentException(s"Missing required option '${PravegaSourceProvider.STREAM_OPTION_KEY}'")
-    }
-  }
-
-  private def buildClientConfig(caseInsensitiveParams: Map[String, String]): ClientConfig = {
+  def buildClientConfig(caseInsensitiveParams: Map[String, String]): ClientConfig = {
     val controllerURI = URI.create(caseInsensitiveParams.getOrElse(PravegaSourceProvider.CONTROLLER_OPTION_KEY, DEFAULT_CONTROLLER))
     ClientConfig.builder()
       .controllerURI(controllerURI)
       .build()
   }
 
-  private def createStreams(caseInsensitiveParams: Map[String, String]): Unit = {
+  def createStream(caseInsensitiveParams: Map[String, String]): Unit = {
     val clientConfig = buildClientConfig(caseInsensitiveParams)
     for (streamManager <- managed(StreamManager.create(clientConfig))) {
       val allowCreateScope = caseInsensitiveParams.getOrElse(PravegaSourceProvider.ALLOW_CREATE_SCOPE_OPTION_KEY, "true").toBoolean
@@ -158,400 +226,6 @@ class PravegaSourceProvider extends DataSourceRegister
     }
   }
 
-  class PravegaMetaTable(metadataTableName: MetadataTableName.MetadataTableName) extends Table with SupportsRead {
-    override def name(): String = "PravegaMataTable"
-
-    override def schema(): StructType = {
-      metadataTableName match {
-        case MetadataTableName.StreamInfo => {
-          StructType(Seq(
-            StructField("head_stream_cut", StringType),
-            StructField("tail_stream_cut", StringType)
-          ))
-        }
-        case MetadataTableName.Streams => {
-          StructType(Seq(
-            StructField("scope_name", StringType),
-            StructField("stream_name", StringType)
-          ))
-        }
-      }
-    }
-
-    override def capabilities(): util.Set[TableCapability] = {
-      import TableCapability._
-      Set(BATCH_READ, BATCH_WRITE, MICRO_BATCH_READ, CONTINUOUS_READ, STREAMING_WRITE,
-        ACCEPT_ANY_SCHEMA).asJava
-    }
-
-    override def newScanBuilder(options: CaseInsensitiveStringMap): ScanBuilder = {
-      () => new PravegaScan(options)
-
-    }
-  }
-
-
-  class PravegaTable extends Table with SupportsWrite with SupportsRead  {
-
-    override def name(): String = "PravegaTable"
-
-    override def schema(): StructType = PravegaReader.pravegaSchema
-
-    override def capabilities(): util.Set[TableCapability] = {
-      import TableCapability._
-      Set(BATCH_READ, BATCH_WRITE, MICRO_BATCH_READ, CONTINUOUS_READ, STREAMING_WRITE,
-        ACCEPT_ANY_SCHEMA).asJava
-    }
-
-    override def newScanBuilder(options: CaseInsensitiveStringMap): ScanBuilder = {
-      () => new PravegaScan(options)
-    }
-
-    override def newWriteBuilder(info: LogicalWriteInfo): WriteBuilder = {
-      new WriteBuilder with SupportsTruncate with SupportsStreamingUpdate {
-        private val inputSchema: StructType = info.schema()
-        private val options = info.options();
-        private val parameters = options.asScala.toMap
-        private val caseInsensitiveParams = parameters.map { case (k, v) => (k.toLowerCase(Locale.ROOT), v) }
-        validateBatchOptions(caseInsensitiveParams)
-        val clientConfig = buildClientConfig(caseInsensitiveParams)
-        val scopeName = caseInsensitiveParams(PravegaSourceProvider.SCOPE_OPTION_KEY)
-        val streamName = caseInsensitiveParams(PravegaSourceProvider.STREAM_OPTION_KEY)
-        val readAfterWriteConsistency = caseInsensitiveParams.getOrElse(PravegaSourceProvider.READ_AFTER_WRITE_CONSISTENCY_OPTION_KEY, "true").toBoolean
-        val transactionStatusPollIntervalMs = caseInsensitiveParams.get(PravegaSourceProvider.TRANSACTION_STATUS_POLL_INTERVAL_MS_OPTION_KEY) match {
-          case Some(s: String) => s.toLong
-          case None => DEFAULT_TRANSACTION_STATUS_POLL_INTERVAL_MS
-        }
-
-        log.info(s"newWriteBuilder: parameters=${parameters}, clientConfig=${clientConfig}")
-
-        createStreams(caseInsensitiveParams)
-
-
-        /**
-         * Creates an optional {@link BatchWrite} to save the data to this data source. Data
-         * sources can return None if there is no writing needed to be done according to the save mode.
-         *
-         * This is used to write a dataframe to a Pravega stream in a batch job.
-         *
-         * If this method fails (by throwing an exception), the action will fail and no Spark job will be
-         * submitted.
-         */
-        override def buildForBatch(): BatchWrite = {
-          assert(inputSchema != null)
-          val transactionTimeoutMs = caseInsensitiveParams.get(PravegaSourceProvider.TRANSACTION_TIMEOUT_MS_OPTION_KEY) match {
-            case Some(s: String) => s.toLong
-            case None => DEFAULT_BATCH_TRANSACTION_TIMEOUT_MS
-          }
-          if (caseInsensitiveParams.getOrElse(PravegaSourceProvider.EXACTLY_ONCE, "true").toBoolean) {
-            new TransactionPravegaWriter(
-              scopeName,
-              streamName,
-              clientConfig,
-              transactionTimeoutMs,
-              readAfterWriteConsistency,
-              transactionStatusPollIntervalMs,
-              inputSchema)
-          } else {
-            new NonTransactionPravegaWriter(
-              scopeName,
-              streamName,
-              clientConfig,
-              inputSchema)
-          }
-
-        }
-
-        /**
-         * Creates an optional {@link StreamingWrite} to save the data to this data source. Data
-         * sources can return None if there is no writing needed to be done.
-         * This is used to write a datastream to a Pravega stream in a structured streaming job.
-         */
-        override def buildForStreaming(): StreamingWrite = {
-          assert(inputSchema != null)
-          val transactionTimeoutMs = caseInsensitiveParams.get(PravegaSourceProvider.TRANSACTION_TIMEOUT_MS_OPTION_KEY) match {
-            case Some(s: String) => s.toLong
-            case None => DEFAULT_TRANSACTION_TIMEOUT_MS
-          }
-          if (caseInsensitiveParams.getOrElse(PravegaSourceProvider.EXACTLY_ONCE, "true").toBoolean) {
-            new TransactionPravegaWriter(
-              scopeName,
-              streamName,
-              clientConfig,
-              transactionTimeoutMs,
-              readAfterWriteConsistency,
-              transactionStatusPollIntervalMs,
-              inputSchema)
-          } else {
-            new NonTransactionPravegaWriter(
-              scopeName,
-              streamName,
-              clientConfig,
-              inputSchema)
-          }
-        }
-
-
-
-        override def truncate(): WriteBuilder = this
-        override def update(): WriteBuilder = this
-      }
-    }
-  }
-
-  class PravegaMetaScan(options: CaseInsensitiveStringMap, metadataTableName: MetadataTableName.MetadataTableName) extends Scan {
-    log.info("in PravegaMetaScan")
-    override def readSchema(): StructType = {
-      metadataTableName match {
-          case MetadataTableName.StreamInfo => {
-            StructType(Seq(
-              StructField("head_stream_cut", StringType),
-              StructField("tail_stream_cut", StringType)
-            ))
-          }
-          case MetadataTableName.Streams => {
-            StructType(Seq(
-              StructField("scope_name", StringType),
-              StructField("stream_name", StringType)
-            ))
-          }
-      }
-    }
-  }
-
-
-
-  class PravegaScan(options: CaseInsensitiveStringMap) extends Scan with Logging {
-
-    override def readSchema(): StructType = PravegaReader.pravegaSchema
-
-    private def createMetadataReader(scopeName: String,
-                                     streamName: String,
-                                     clientConfig: ClientConfig,
-                                     metadataTableName: MetadataTableName.MetadataTableName,
-                                     caseInsensitiveParams: Map[String, String]): Batch = {
-      (for (streamManager <- managed(StreamManager.create(clientConfig))) yield {
-        metadataTableName match {
-          case MetadataTableName.StreamInfo => {
-            val streamInfo = streamManager.getStreamInfo(scopeName, streamName)
-            val schema = StructType(Seq(
-              StructField("head_stream_cut", StringType),
-              StructField("tail_stream_cut", StringType)
-            ))
-            val row = new GenericInternalRow(Seq(
-              UTF8String.fromString(streamInfo.getHeadStreamCut.asText()),
-              UTF8String.fromString(streamInfo.getTailStreamCut.asText())
-            ).toArray[Any])
-            MemoryDataSourceReader(schema, Seq(row))
-          }
-          case MetadataTableName.Streams => {
-            val streams = streamManager.listStreams(scopeName)
-            val schema = StructType(Seq(
-              StructField("scope_name", StringType),
-              StructField("stream_name", StringType)
-            ))
-
-            var rows = Seq[InternalRow]()
-
-            streams.asScala.foreach(stream => {
-              rows = rows :+ new GenericInternalRow(
-                Seq(
-                  UTF8String.fromString(stream.getScope),
-                  UTF8String.fromString(stream.getStreamName)
-                ).toArray[Any])
-            })
-
-            MemoryDataSourceReader(schema, rows)
-          }
-        }
-      }).acquireAndGet(identity _)
-    }
-
-
-    /**
-     * Creates a {@link Batch} to scan the data from this data source.
-     *
-     * This is used to read a Pravega stream as a dataframe in a batch job.
-     *
-     * Unbounded stream cuts (earliest, latest, unbounded) are bound only once.
-     * Late binding is not available.
-     *
-     * If this method fails (by throwing an exception), the action will fail and no Spark job will be
-     * submitted.
-     *
-     */
-    override def toBatch(): Batch = {
-
-      val caseInsensitiveParams = CaseInsensitiveMap(options.asScala.toMap)
-      validateBatchOptions(caseInsensitiveParams)
-      val clientConfig = buildClientConfig(caseInsensitiveParams)
-      val scopeName = caseInsensitiveParams(PravegaSourceProvider.SCOPE_OPTION_KEY)
-      val streamName = caseInsensitiveParams(PravegaSourceProvider.STREAM_OPTION_KEY)
-
-      val metadataTableName = caseInsensitiveParams.get(PravegaSourceProvider.METADATA_OPTION_KEY).map(MetadataTableName.withName)
-      if (metadataTableName.isDefined) {
-        log.info("in metadataTableName")
-        return createMetadataReader(
-          scopeName,
-          streamName,
-          clientConfig,
-          metadataTableName.get,
-          caseInsensitiveParams)
-      }
-
-      val startStreamCut = PravegaSourceProvider.getPravegaStreamCut(
-        caseInsensitiveParams, PravegaSourceProvider.START_STREAM_CUT_OPTION_KEY, EarliestStreamCut)
-
-      val endStreamCut = PravegaSourceProvider.getPravegaStreamCut(
-        caseInsensitiveParams, PravegaSourceProvider.END_STREAM_CUT_OPTION_KEY, LatestStreamCut)
-
-      log.info(s"toBatch: clientConfig=${clientConfig}, scopeName=${scopeName}, streamName=${streamName}"
-        + s" startStreamCut=${startStreamCut}, endStreamCut=${endStreamCut}")
-
-      createStreams(caseInsensitiveParams)
-
-      new PravegaBatch(
-        scopeName,
-        streamName,
-        clientConfig,
-        caseInsensitiveParams,
-        startStreamCut,
-        endStreamCut)
-    }
-
-
-    /**
-     * Creates a {@link MicroBatchStream} to read batches of data from this data source in a
-     * streaming query.
-     * This is used to read a Pravega stream as a datastream in a structured streaming job.
-     *
-     * The execution engine will create a micro-batch reader at the start of a streaming query,
-     * alternate calls to setOffsetRange and planInputPartitions for each batch to process, and
-     * then call stop() when the execution is complete. Note that a single query may have multiple
-     * executions due to restart or failure recovery.
-     *
-     * @param checkpointLocation a path to Hadoop FS scratch space that can be used for failure
-     *                           recovery. Readers for the same logical source in the same query
-     *                           will be given the same checkpointLocation. The Pravega connector
-     *                           does not need to implement this checkpointLocation since the spark
-     *                           handles the Pravega streamcut through the Offset interface.
-     */
-    override def toMicroBatchStream(checkpointLocation: String): MicroBatchStream = {
-      val caseInsensitiveParams = CaseInsensitiveMap(options.asScala.toMap)
-      validateStreamOptions(caseInsensitiveParams)
-
-      val clientConfig = buildClientConfig(caseInsensitiveParams)
-      val scopeName = caseInsensitiveParams(PravegaSourceProvider.SCOPE_OPTION_KEY)
-      val streamName = caseInsensitiveParams(PravegaSourceProvider.STREAM_OPTION_KEY)
-
-      val startStreamCut = PravegaSourceProvider.getPravegaStreamCut(
-      caseInsensitiveParams, PravegaSourceProvider.START_STREAM_CUT_OPTION_KEY, LatestStreamCut)
-
-      val endStreamCut = PravegaSourceProvider.getPravegaStreamCut(
-      caseInsensitiveParams, PravegaSourceProvider.END_STREAM_CUT_OPTION_KEY, UnboundedStreamCut)
-
-      log.info(s"toMicroBatchStream: clientConfig=${clientConfig}, scopeName=${scopeName}, streamName=${streamName}"
-      + s" startStreamCut=${startStreamCut}, endStreamCut=${endStreamCut}")
-
-      createStreams(caseInsensitiveParams)
-
-      new PravegaMicroBatchStream(
-        scopeName,
-        streamName,
-        clientConfig,
-        caseInsensitiveParams,
-        startStreamCut,
-        endStreamCut)
-    }
-
-  }
-
-}
-
-
-object PravegaSourceProvider extends Logging {
-  private[spark] val SOURCE_PROVIDER_NAME = "pravega"
-  private[spark] val CONTROLLER_OPTION_KEY = "controller"
-  private[spark] val SCOPE_OPTION_KEY = "scope"
-  private[spark] val STREAM_OPTION_KEY = "stream"
-  private[spark] val TRANSACTION_TIMEOUT_MS_OPTION_KEY = "transaction_timeout_ms"
-  private[spark] val START_STREAM_CUT_OPTION_KEY = "start_stream_cut"
-  private[spark] val END_STREAM_CUT_OPTION_KEY = "end_stream_cut"
-  private[spark] val ALLOW_CREATE_SCOPE_OPTION_KEY = "allow_create_scope"
-  private[spark] val ALLOW_CREATE_STREAM_OPTION_KEY = "allow_create_stream"
-  private[spark] val DEFAULT_NUM_SEGMENTS_OPTION_KEY = "default_num_segments"
-  private[spark] val READ_AFTER_WRITE_CONSISTENCY_OPTION_KEY = "read_after_write_consistency"
-  private[spark] val TRANSACTION_STATUS_POLL_INTERVAL_MS_OPTION_KEY = "transaction_status_poll_interval_ms"
-  private[spark] val METADATA_OPTION_KEY = "metadata"
-  private[spark] val DEFAULT_SCALE_FACTOR_OPTION_KEY = "default_scale_factor"
-  private[spark] val DEFAULT_SEGMENT_TARGET_RATE_BYTES_PER_SEC_OPTION_KEY = "default_segment_target_rate_bytes_per_sec"
-  private[spark] val DEFAULT_SEGMENT_TARGET_RATE_EVENTS_PER_SEC_OPTION_KEY = "default_segment_target_rate_events_per_sec"
-  private[spark] val DEFAULT_RETENTION_DURATION_MILLISECONDS_OPTION_KEY = "default_retention_duration_milliseconds"
-  private[spark] val DEFAULT_RETENTION_SIZE_BYTES_OPTION_KEY = "default_retention_size_bytes"
-  private[spark] val EXACTLY_ONCE = "exactly_once"
-
-  private[spark] val STREAM_CUT_EARLIEST = "earliest"
-  private[spark] val STREAM_CUT_LATEST = "latest"
-  private[spark] val STREAM_CUT_UNBOUNDED = "unbounded"
-  private[spark] val ROUTING_KEY_ATTRIBUTE_NAME = "routing_key"
-  private[spark] val EVENT_ATTRIBUTE_NAME = "event"
-
-
-  def buildStreamConfig(caseInsensitiveParams: Map[String, String]): StreamConfiguration = {
-    var streamConfig = StreamConfiguration.builder
-    val minSegments = caseInsensitiveParams.get(PravegaSourceProvider.DEFAULT_NUM_SEGMENTS_OPTION_KEY)
-    val scaleFactor = caseInsensitiveParams.get(PravegaSourceProvider.DEFAULT_SCALE_FACTOR_OPTION_KEY)
-    val targetRateBytesPerSec = caseInsensitiveParams.get(PravegaSourceProvider.DEFAULT_SEGMENT_TARGET_RATE_BYTES_PER_SEC_OPTION_KEY)
-    val targetRateEventsPerSec = caseInsensitiveParams.get(PravegaSourceProvider.DEFAULT_SEGMENT_TARGET_RATE_EVENTS_PER_SEC_OPTION_KEY)
-    val retentionDurationMilliseconds = caseInsensitiveParams.get(PravegaSourceProvider.DEFAULT_RETENTION_DURATION_MILLISECONDS_OPTION_KEY)
-    val retentionSizeBytes = caseInsensitiveParams.get(PravegaSourceProvider.DEFAULT_RETENTION_SIZE_BYTES_OPTION_KEY)
-
-    // set scaling policy
-    streamConfig = (minSegments, scaleFactor, targetRateBytesPerSec, targetRateEventsPerSec) match {
-      case (Some(minSegments), scaleFactor, targetRateKiloBytesPerSec, targetRateEventsPerSec) =>
-        (scaleFactor, targetRateKiloBytesPerSec, targetRateEventsPerSec) match {
-          case (Some(scaleFactor), Some(targetRateBytesPerSec), None) =>
-            streamConfig.scalingPolicy(ScalingPolicy.byDataRate(targetRateBytesPerSec.toInt / 1024, scaleFactor.toInt, minSegments.toInt))
-          case (Some(scaleFactor), None, Some(targetRateEventsPerSec)) =>
-            streamConfig.scalingPolicy(ScalingPolicy.byEventRate(targetRateEventsPerSec.toInt, scaleFactor.toInt, minSegments.toInt))
-          case _ =>
-            streamConfig.scalingPolicy(ScalingPolicy.fixed(minSegments.toInt))
-        }
-      case _ => streamConfig
-    }
-
-    // set retention policy
-    streamConfig = (retentionDurationMilliseconds, retentionSizeBytes) match {
-      case (Some(retentionDurationMilliseconds), None) =>
-        streamConfig.retentionPolicy(RetentionPolicy.byTime(Duration.ofMillis(retentionDurationMilliseconds.toLong)))
-      case (None, Some(retentionSizeBytes)) =>
-        streamConfig.retentionPolicy(RetentionPolicy.bySizeBytes(retentionSizeBytes.toLong))
-      case _ =>
-        streamConfig
-    }
-
-    log.info("streamConfig is {}", streamConfig)
-
-
-    return streamConfig.build()
-  }
-
-  def getPravegaStreamCut(
-                           params: Map[String, String],
-                           streamCutOptionKey: String,
-                           defaultStreamCut: PravegaStreamCut): PravegaStreamCut = {
-
-    params.get(streamCutOptionKey).map(_.trim) match {
-      case Some(offset) if offset.toLowerCase(Locale.ROOT) == STREAM_CUT_LATEST =>
-        LatestStreamCut
-      case Some(offset) if offset.toLowerCase(Locale.ROOT) == STREAM_CUT_EARLIEST =>
-        EarliestStreamCut
-      case Some(offset) if offset.toLowerCase(Locale.ROOT) == STREAM_CUT_UNBOUNDED =>
-        UnboundedStreamCut
-      case Some(base64String) => SpecificStreamCut(StreamCut.from(base64String))
-      case None => defaultStreamCut
-    }
-  }
 }
 
 object PravegaReader {
